LayoutTests/storage/filesystemaccess/file-handle-getfile-worker-expected.txt

[Worker] This test checks getFile() of FileSystemFileHandle.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Starting worker: resources/file-handle-getfile.js
PASS [Worker] writeSize is writeBuffer.byteLength
PASS [Worker] readText is fileContent
PASS successfullyParsed is true

TEST COMPLETE
